Move useWindowSize from utils to hooks.ts
