+++ lib/hooks.ts
@@ -1,0 +1,21 @@
+import {useEffect, useState} from "react";
+
+const useWindowSize = () => {
+  const [windowSize, setWindowSize] = useState({
+    width: 0,
+    height: 0
+  });
+  useEffect(() => {
+    if (typeof window === 'undefined') return;
+    const onResize = () => {
+      setWindowSize({
+        width: window.innerWidth,
+        height: window.innerHeight
+      });
+    }
+    window.addEventListener('resize', onResize);
+    onResize();
+    return () => window.removeEventListener('resize', onResize);
+  }, []);
+  return windowSize;
+}
